refactor: uncouple push context definition from push contact service
